consensus/bor: fix build

+++ consensus/bor/statefull/processor.go
@@ -181,7 +181,7 @@ func ApplyStateSyncEvents(vmenv *vm.EVM, tx *types.Transaction, message *core.Me
 		// it here, we use the block time. The calldata will be different than what
 		// was constructed while executing the transaction but it'll be deterministic
 		// in every run.
-		now = vmenv.Context.Time
+		now = int64(vmenv.Context.Time)
 	)
 
 	for _, event := range events {
